pyproject.toml
[project]
name = "transfer-market-game"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "bcrypt>=5.0.0",
    "beautifulsoup4>=4.14.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.50.0",
]
